requirements.txt
Flask==2.1.3
Werkzeug==2.1.2
requests==2.31.0
python-dotenv==1.0.0
customtkinter=0.3
Pillow
